Create all handlers
